requirements.txt
# Core Python and Data Science
pandas>=1.4.0
scikit-learn>=1.1.0
numpy>=1.22.0
python-dotenv
# Gradient Boosting Libraries
xgboost>=1.6.0
lightgbm>=3.3.0
catboost>=1.0.6

# Imbalanced learning
imbalanced-learn>=0.9.0

# Optimization
optuna>=3.0.0

# Experiment Tracking
mlflow>=2.0.0

# CLI App Framework
typer[all]>=0.7.0

# Optional but Recommended for Plotting (if using in notebooks)
matplotlib>=3.5.0
seaborn>=0.11.2
plotly>=5.6.0
